requirements.txt
Django==2.1.5
djangorestframework
markdown
django-filter
pytz==2018.9
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
django-rest-swagger